src/sqlagent/requirements.txt
langchain
langchain-openai
python-dotenv
langchain-community
pydantic